moved AssetSystemStub to TestUtils folder
