+++ Gems/Atom/Utils/Code/Include/Atom/Utils/TestUtils/AssetSystemStub.h
