+++ Gems/Atom/Utils/Code/Source/TestUtils/AssetSystemStub.cpp
@@ -6,7 +6,7 @@
  *
  */
 
-#include <Atom/Utils/AssetSystemStub.h>
+#include <Atom/Utils/TestUtils/AssetSystemStub.h>
 #include <AzFramework/StringFunc/StringFunc.h>
 
 namespace UnitTest
